adjust tresholds

+++ c/3-pid/controller.h
@@ -29,10 +29,10 @@ typedef struct {
   uint32_t reads_per_bin;
   /// When ADC reads below this signal, the state is set to `close` to the
   /// motor magnet. If the state has changed, a new revolution is counted.
-  uint8_t revolution_treshold_close;
+  float revolution_treshold_close;
   /// When ADC reads above this signal, the state is set to `far` from the
   /// motor magnet.
-  uint8_t revolution_treshold_far;
+  float revolution_treshold_far;
   /// Linux PWM channel to use.
   uint8_t pwm_channel;
   /// Frequency of the PWM signal.
